question and commit

+++ src/main/java/school/community/controller/QuestionController.java
@@ -22,7 +22,7 @@ public class QuestionController {
 
 
     @GetMapping("/question/{id}")
-    public  String question(@PathVariable(name="id") Integer id, Model model){
+    public  String question(@PathVariable(name="id") Long id, Model model){
 
         QuestionDTO questionDTO=questionService.getById(id);
         //累加阅读数
